requirements.txt
# Core dependencies
python-dotenv>=0.19.0
pydantic>=2.11.0
langchain>=0.3.0
langchain-groq>=0.3.0
langchain-core>=0.3.0

# Database and embeddings
chromadb>=1.0.0
sentence-transformers>=2.2.0

# Research APIs
arxiv>=2.2.0
semanticscholar>=0.10.0

# Async support
aiofiles>=0.8.0

# Other dependencies
requests>=2.32.0

# Optional: For CLI and colored output (if you use them)
colorama>=0.4.6

# For scientific computing (if you use pandas, numpy, scikit-learn, etc.)
numpy>=1.26.0
scikit-learn>=1.5.0

# For plotting (if you use matplotlib, seaborn, etc.)
matplotlib>=3.9.0
seaborn>=0.13.0

# For pandas (if you use it)
pandas>=2.2.0